backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
pandas==2.2.2
langdetect==1.0.9
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.5.4
pymongo==4.6.1
gunicorn==21.2.0
